refactor(KvWwwHeaderBasic): move header utils to src/utils with standard names
Relocate the header-local utils into the package's standard src/utils dir
and rename for clarity:
- searchEngine.ts  -> typeaheadSearchEngine.ts
- searchConfig.ts  -> typeaheadSearchConfig.ts
- navLinks.ts      -> headerNavLinks.ts

Move searchEngine.spec.ts to tests/unit/specs/utils/ as
typeaheadSearchEngine.spec.ts. Update all imports to the #utils/ alias.
