+++ @kiva/kv-components/src/utils/typeaheadSearchEngine.ts
